clean up
Discovered that Riot API limits the number of calls to certain APIs,
thus preventing us from further being able to test the program and play
around with numbers. We really cannot procced for now as we cannot know
what we are supposed to read without the API allowing us to call it.